更新：修改customerTest的bug，删除instrument details

+++ src/main/java/team/wy/orchestra/biz/CustomerBiz.java
@@ -15,7 +15,7 @@ public class CustomerBiz {
      * @return Customer
      */
 
-    public Customer getCustomerByid(int id) {
+    public Customer getCustomerByid(long id) {
         try {
             return CustomerDao.getCustomerByid(id);
         } catch (SQLException e) {
